+++ pkg/custom/doc.go
@@ -1,4 +1,4 @@
 /*
-Package custom contains definitions of Capillaries custom processors
+Package custom contains definitions of Capillaries custom processors. Each processor must implement CustomProcessorRunner interface.
 */
 package custom
