+++ pkg/eval/doc.go
@@ -1,4 +1,5 @@
 /*
-Package eval contains base implementation of the AST expression eval engine with aggregation capabilities
+Package eval contains base implementation of the AST expression eval engine with aggregation capabilities.
+The goal is to keep this package as generic as posible, as it is being used by Capillaries processors and gocqlmem Cassandra-like engine.
 */
 package eval
