+++ pkg/exe/toolbelt/capitoolbelt.go
@@ -661,7 +661,7 @@ func protoFileReaderCreator() int {
 
 // Used by Toolbelt (exec_node command), can be used for testing Capillaries node execution without message queue wokflow
 func runNode(envConfig *env.EnvConfig, logger *l.CapiLogger, nodeName string, runId int16, scriptFilePath string, paramsFilePath string, cqlSession gocqlshims.Session, keyspace string) (int16, error) {
-	logger.PushF("api.RunNode")
+	logger.PushF("toolbelt.runNode")
 	defer logger.PopF()
 
 	script, _, err := sc.NewScriptFromFiles(envConfig.CaPath, envConfig.PrivateKeys, scriptFilePath, paramsFilePath, envConfig.CustomProcessorDefFactoryInstance, envConfig.CustomProcessorsSettings)
