+++ pkg/proc/proc_table_creator_lookup.go
@@ -1,12 +1,8 @@
 package proc
 
 import (
-	"bufio"
 	"errors"
 	"fmt"
-	"io"
-	"net/url"
-	"os"
 	"time"
 
 	"github.com/capillariesio/capillaries/pkg/cql"
@@ -16,602 +12,9 @@ import (
 	"github.com/capillariesio/capillaries/pkg/evalcapi"
 	"github.com/capillariesio/capillaries/pkg/l"
 	"github.com/capillariesio/capillaries/pkg/sc"
-	"github.com/capillariesio/capillaries/pkg/xfer"
 )
 
-type TableRecord map[string]any
-type TableRecordPtr *map[string]any
-type TableRecordBatch []TableRecordPtr
-
-// const ReadFileTableInserterBatchSize int = 500 // 1000 still may cause OOM on c7g.large daemon boxes when loading portfolio bigtest parquet files
-// const TableToTableInserterBatchSize int = 500  // Same as above, although there were no OOMs reported yet
 const MaxAmazonKeyspacesInElements int = 100
-
-/*
-func reportWriteTable(logger *l.CapiLogger, pCtx *ctx.MessageProcessingContext, recordCount int, dur time.Duration, indexCount int, workerCount int) {
-	logger.InfoCtx(pCtx, "WriteTable: %d items in %.3fs (%.0f items/s, %d indexes, eff rate %.0f writes/s), %d workers",
-		recordCount,
-		dur.Seconds(),
-		float64(recordCount)/dur.Seconds(),
-		indexCount,
-		float64(recordCount*(indexCount+1))/dur.Seconds(),
-		workerCount)
-}
-*/
-
-/*
-func reportWriteTableLeftovers(logger *l.CapiLogger, pCtx *ctx.MessageProcessingContext, recordCount int, dur time.Duration, indexCount int, workerCount int) {
-	logger.InfoCtx(pCtx, "WriteTableLeftovers: %d items in %.3fs (%.0f items/s, %d indexes, eff rate %.0f writes/s), %d workers",
-		recordCount,
-		dur.Seconds(),
-		float64(recordCount)/dur.Seconds(),
-		indexCount,
-		float64(recordCount*(indexCount+1))/dur.Seconds(),
-		workerCount)
-}
-*/
-
-func reportWriteTableComplete(logger *l.CapiLogger, pCtx *ctx.MessageProcessingContext, readCount int, recordCount int, dur time.Duration, indexCount int, workerCount int) {
-	logger.InfoCtx(pCtx, "WriteTableComplete: read %d, wrote %d items in %.3fs (%.0f items/s, %d indexes, eff rate %.0f writes/s), %d workers",
-		readCount,
-		recordCount,
-		dur.Seconds(),
-		float64(recordCount)/dur.Seconds(),
-		indexCount,
-		float64(recordCount*(indexCount+1))/dur.Seconds(),
-		workerCount)
-}
-
-func checkRunReadFileForBatchSanity(node *sc.ScriptNodeDef, srcFileIdx int) error {
-	if !node.HasFileReader() {
-		return errors.New("node does not have file reader")
-	}
-	if !node.HasTableCreator() {
-		return errors.New("node does not have table creator")
-	}
-
-	if srcFileIdx < 0 || srcFileIdx >= len(node.FileReader.SrcFileUrls) {
-		return fmt.Errorf("cannot find file to read: asked to read src file with index %d while there are only %d source files available", srcFileIdx, len(node.FileReader.SrcFileUrls))
-	}
-	return nil
-}
-
-func RunReadFileForBatch(envConfig *env.EnvConfig, logger *l.CapiLogger, pCtx *ctx.MessageProcessingContext, srcFileIdx int) (BatchStats, error) {
-	logger.PushF("proc.RunReadFileForBatch")
-	defer logger.PopF()
-
-	totalStartTime := time.Now()
-	bs := BatchStats{RowsRead: 0, RowsWritten: 0}
-
-	node := pCtx.CurrentScriptNode
-
-	if err := checkRunReadFileForBatchSanity(node, srcFileIdx); err != nil {
-		return bs, err
-	}
-
-	filePath := node.FileReader.SrcFileUrls[srcFileIdx]
-
-	u, err := url.Parse(filePath)
-	if err != nil {
-		return bs, fmt.Errorf("cannot parse file url %s: %s", filePath, err.Error())
-	}
-
-	bs.Src = filePath
-	bs.Dst = node.TableCreator.Name + cql.RunIdSuffix(pCtx.Msg.RunId)
-
-	var localSrcFile *os.File
-	var fileReader io.Reader
-	var fileReadSeeker io.ReadSeeker
-	if u.Scheme == xfer.UrlSchemeFile || len(u.Scheme) == 0 {
-		localSrcFile, err = os.Open(filePath)
-		if err != nil {
-			return bs, err
-		}
-		defer localSrcFile.Close()
-		fileReader = bufio.NewReader(localSrcFile)
-		fileReadSeeker = localSrcFile
-	} else if u.Scheme == xfer.UrlSchemeHttp || u.Scheme == xfer.UrlSchemeHttps || u.Scheme == xfer.UrlSchemeS3 {
-		var readCloser io.ReadCloser
-		switch u.Scheme {
-		case xfer.UrlSchemeHttp, xfer.UrlSchemeHttps:
-			readCloser, err = xfer.GetHttpReadCloser(filePath, u.Scheme, envConfig.CaPath)
-			if err != nil {
-				return bs, fmt.Errorf("cannot open http file %s: %s", filePath, err.Error())
-			}
-		case xfer.UrlSchemeS3:
-			readCloser, err = xfer.GetS3ReadCloser(filePath)
-			if err != nil {
-				return bs, fmt.Errorf("cannot open s3 file %s: %s", filePath, err.Error())
-			}
-		default:
-			return bs, fmt.Errorf("cannot open file %s: unknown url scheme", filePath)
-		}
-		defer readCloser.Close()
-
-		// If this is a parquet file, download it and then open so we have fileReadSeeker
-		if node.FileReader.ReaderFileType == sc.ReaderFileTypeParquet {
-			dstFile, err := os.CreateTemp("", "capi")
-			if err != nil {
-				return bs, fmt.Errorf("cannot create temp file for %s: %s", filePath, err.Error())
-			}
-
-			if _, err := io.Copy(dstFile, readCloser); err != nil {
-				dstFile.Close()
-				return bs, fmt.Errorf("cannot save http file %s to temp file %s: %s", filePath, dstFile.Name(), err.Error())
-			}
-
-			logger.Info("downloaded http file %s to %s", filePath, dstFile.Name())
-			dstFile.Close()
-			defer os.Remove(dstFile.Name())
-
-			localSrcFile, err = os.Open(dstFile.Name())
-			if err != nil {
-				return bs, fmt.Errorf("cannot read from file %s downloaded from %s: %s", dstFile.Name(), filePath, err.Error())
-			}
-			defer localSrcFile.Close()
-			fileReadSeeker = localSrcFile
-		} else {
-			// Just read from the net
-			fileReader = readCloser
-			defer readCloser.Close()
-		}
-	} else if u.Scheme == xfer.UrlSchemeSftp {
-		// When dealing with sftp, we download the *whole* file, instead of providing a reader
-		dstFile, err := os.CreateTemp("", "capi")
-		if err != nil {
-			return bs, fmt.Errorf("cannot create temp file for %s: %s", filePath, err.Error())
-		}
-
-		// Download and schedule delete
-		if err = xfer.DownloadSftpFile(filePath, envConfig.PrivateKeys, dstFile); err != nil {
-			dstFile.Close()
-			return bs, err
-		}
-		logger.Info("downloaded sftp file %s to %s", filePath, dstFile.Name())
-		dstFile.Close()
-		defer os.Remove(dstFile.Name())
-
-		// Create a reader for the temp file
-		localSrcFile, err = os.Open(dstFile.Name())
-		if err != nil {
-			return bs, fmt.Errorf("cannot read from file %s downloaded from %s: %s", dstFile.Name(), filePath, err.Error())
-		}
-		defer localSrcFile.Close()
-		fileReader = bufio.NewReader(localSrcFile)
-		fileReadSeeker = localSrcFile
-	} else {
-		return bs, fmt.Errorf("l scheme %s not supported: %s", u.Scheme, filePath)
-	}
-
-	switch node.FileReader.ReaderFileType {
-	case sc.ReaderFileTypeCsv:
-		return readCsv(envConfig, logger, pCtx, totalStartTime, filePath, fileReader)
-	case sc.ReaderFileTypeParquet:
-		return readParquet(envConfig, logger, pCtx, totalStartTime, filePath, fileReadSeeker)
-	default:
-		return BatchStats{RowsRead: 0, RowsWritten: 0}, fmt.Errorf("unknown reader file type: %d", node.FileReader.ReaderFileType)
-	}
-}
-
-func RunCreateTableForCustomProcessorForBatch(envConfig *env.EnvConfig,
-	logger *l.CapiLogger,
-	pCtx *ctx.MessageProcessingContext,
-	readerNodeRunId int16,
-	startLeftToken int64,
-	endLeftToken int64) (BatchStats, error) {
-
-	logger.PushF("proc.RunCreateTableForCustomProcessorForBatch")
-	defer logger.PopF()
-
-	node := pCtx.CurrentScriptNode
-
-	totalStartTime := time.Now()
-	bs := BatchStats{RowsRead: 0, RowsWritten: 0, Src: node.TableReader.TableName + cql.RunIdSuffix(readerNodeRunId), Dst: node.TableCreator.Name + cql.RunIdSuffix(readerNodeRunId)}
-
-	if readerNodeRunId == 0 {
-		return bs, errors.New("this node has a dependency node to read data from that was never started in this keyspace (readerNodeRunId == 0)")
-	}
-
-	if !node.HasTableReader() {
-		return bs, errors.New("node does not have table reader")
-	}
-	if !node.HasTableCreator() {
-		return bs, errors.New("node does not have table creator")
-	}
-
-	// Fields to read from source table
-	srcLeftFieldRefs := sc.FieldRefs{}
-	srcLeftFieldRefs.AppendWithFilter(*node.CustomProcessor.GetUsedInTargetExpressionsFields(), sc.ReaderAlias)
-	srcLeftFieldRefs.AppendWithFilter(node.TableCreator.UsedInTargetExpressionsFields, sc.ReaderAlias)
-
-	leftBatchSize := node.TableReader.RowsetSize
-	curStartLeftToken := startLeftToken
-
-	rsIn := NewRowsetFromFieldRefs(
-		sc.FieldRefs{sc.RowidFieldRef(node.TableReader.TableName)},
-		sc.FieldRefs{sc.RowidTokenFieldRef()},
-		srcLeftFieldRefs)
-
-	instr, err := createInserterAndStartWorkers(logger, envConfig, pCtx, &node.TableCreator, DataIdxSeqModeDataFirst, logger.ZapMachine.String)
-	if err != nil {
-		return bs, err
-	}
-	defer instr.closeInserter(logger, pCtx)
-
-	flushVarsArrayCallback := func(varsArray []eval.VarValuesMap, varsArrayCount int) error {
-
-		instr.startDrainer()
-
-		// Minimize allocations to help GC in this high-traffic loop
-		var tableRecord map[string]any
-		indexKeyMap := map[string]string{}
-		var inResult bool
-		var err error
-		for outRowIdx := 0; outRowIdx < varsArrayCount; outRowIdx++ {
-			vars := varsArray[outRowIdx]
-
-			tableRecord, err = node.TableCreator.CalculateTableRecordFromSrcVars(vars)
-			if err != nil {
-				instr.cancelDrainer(fmt.Errorf("cannot populate table record from [%v], node %s: [%s]", vars, node.Name, err.Error()))
-				return instr.waitForDrainer()
-			}
-
-			// Check table creator having
-			inResult, err = node.TableCreator.CheckTableRecordHavingCondition(tableRecord)
-			if err != nil {
-				instr.cancelDrainer(fmt.Errorf("cannot check having condition [%s], node %s, table record [%v]: [%s]", node.TableCreator.RawHaving, node.Name, tableRecord, err.Error()))
-				return instr.waitForDrainer()
-			}
-
-			if inResult {
-				err = instr.buildIndexKeys(tableRecord, indexKeyMap)
-				if err != nil {
-					instr.cancelDrainer(fmt.Errorf("cannot build index keys for table %s: [%s]", node.TableCreator.Name, err.Error()))
-					return instr.waitForDrainer()
-				}
-
-				instr.add(tableRecord, indexKeyMap)
-				bs.RowsWritten++
-			}
-		}
-
-		instr.doneSending()
-		return instr.waitForDrainer()
-	}
-
-	var curStartLeftTokenRowIds []int64
-	for {
-		lastRetrievedLeftToken, endTokenRowIds, err := selectBatchFromTableByToken(logger,
-			pCtx,
-			rsIn,
-			node.TableReader.TableName,
-			readerNodeRunId,
-			leftBatchSize,
-			curStartLeftToken,
-			endLeftToken,
-			curStartLeftTokenRowIds)
-		if err != nil {
-			return bs, err
-		}
-
-		// If token(rowid) guaranteed uniqueness, we would just "curStartLeftToken = lastRetrievedLeftToken + 1"
-		// But duplicates are possible, so we have to be prepared to handle token overlaps
-		// (rows with same token but different rowids returned in separate selectBatchFromTableByToken calls)
-		// See overlap/epilogue logic in selectBatchFromTableByToken.
-		curStartLeftToken = lastRetrievedLeftToken
-		curStartLeftTokenRowIds = endTokenRowIds
-
-		if rsIn.RowCount == 0 {
-			break
-		}
-		customProcBatchStartTime := time.Now()
-
-		if err = node.CustomProcessor.(CustomProcessorRunner).Run(logger, pCtx, rsIn, flushVarsArrayCallback); err != nil {
-			return bs, err
-		}
-
-		custProcDur := time.Since(customProcBatchStartTime)
-		logger.InfoCtx(pCtx, "CustomProcessor: %d items in %v (%.0f items/s)", rsIn.RowCount, custProcDur, float64(rsIn.RowCount)/custProcDur.Seconds())
-
-		// The frequency of this hearbeat may not be enough, even for small rowset_size:
-		// Python calculations even for one row may take, say, a minute
-		// We cannot make heartbeats more granular here (each rowset is handled as a single Python command),
-		// so our last resort is probably to increase the dead hearbeat timeout in CapiMQ
-		instr.PCtx.SendHeartbeat()
-
-		bs.RowsRead += rsIn.RowCount
-
-		// We are tempted to "if rs.RowCount < srcBatchSize break", here but do not do that:
-		// because of the rowid overlapping/epilogue logic, selectBatchFromTableByToken returns less rows than rs capacity
-
-	} // for each source table batch
-
-	bs.Elapsed = time.Since(totalStartTime)
-	reportWriteTableComplete(logger, pCtx, bs.RowsRead, bs.RowsWritten, bs.Elapsed, len(node.TableCreator.Indexes), instr.NumWorkers)
-
-	return bs, nil
-}
-
-func RunCreateTableForBatch(envConfig *env.EnvConfig,
-	logger *l.CapiLogger,
-	pCtx *ctx.MessageProcessingContext,
-	readerNodeRunId int16,
-	startLeftToken int64,
-	endLeftToken int64) (BatchStats, error) {
-
-	logger.PushF("proc.RunCreateTableForBatch")
-	defer logger.PopF()
-
-	node := pCtx.CurrentScriptNode
-
-	// batchStartTime := time.Now()
-	totalStartTime := time.Now()
-	bs := BatchStats{RowsRead: 0, RowsWritten: 0, Src: node.TableReader.TableName + cql.RunIdSuffix(readerNodeRunId), Dst: node.TableCreator.Name + cql.RunIdSuffix(readerNodeRunId)}
-
-	if readerNodeRunId == 0 {
-		return bs, errors.New("this node has a dependency node to read data from that was never started in this keyspace (readerNodeRunId == 0)")
-	}
-
-	if !node.HasTableReader() {
-		return bs, errors.New("node does not have table reader")
-	}
-	if !node.HasTableCreator() {
-		return bs, errors.New("node does not have table creator")
-	}
-
-	// Fields to read from source table
-	srcLeftFieldRefs := sc.FieldRefs{}
-	srcLeftFieldRefs.AppendWithFilter(node.TableCreator.UsedInTargetExpressionsFields, sc.ReaderAlias)
-
-	leftBatchSize := node.TableReader.RowsetSize
-	// tableRecordBatchCount := 0
-	curStartLeftToken := startLeftToken
-
-	rsIn := NewRowsetFromFieldRefs(
-		sc.FieldRefs{sc.RowidFieldRef(node.TableReader.TableName)},
-		sc.FieldRefs{sc.RowidTokenFieldRef()},
-		srcLeftFieldRefs)
-
-	instr, err := createInserterAndStartWorkers(logger, envConfig, pCtx, &node.TableCreator, DataIdxSeqModeDataFirst, logger.ZapMachine.String)
-	if err != nil {
-		return bs, err
-	}
-	instr.startDrainer()
-	defer instr.closeInserter(logger, pCtx)
-
-	var curStartLeftTokenRowIds []int64
-	for {
-		lastRetrievedLeftToken, endTokenRowIds, err := selectBatchFromTableByToken(logger,
-			pCtx,
-			rsIn,
-			node.TableReader.TableName,
-			readerNodeRunId,
-			leftBatchSize,
-			curStartLeftToken,
-			endLeftToken,
-			curStartLeftTokenRowIds)
-		if err != nil {
-			instr.cancelDrainer(fmt.Errorf("cannot select batch from source table, node %s: %s", node.Name, err.Error()))
-			return bs, instr.waitForDrainer()
-		}
-
-		// If token(rowid) guaranteed uniqueness, we would just "curStartLeftToken = lastRetrievedLeftToken + 1"
-		// But duplicates are possible, so we have to be prepared to handle token overlaps
-		// (rows with same token but different rowids returned in separate selectBatchFromTableByToken calls)
-		// See overlap/epilogue logic in selectBatchFromTableByToken.
-		curStartLeftToken = lastRetrievedLeftToken
-		curStartLeftTokenRowIds = endTokenRowIds
-
-		if rsIn.RowCount == 0 {
-			break
-		}
-
-		// Minimize allocations to help GC in this high-traffic loop
-		var tableRecord map[string]any
-		indexKeyMap := map[string]string{}
-		vars := eval.VarValuesMap{}
-		var inResult bool
-
-		// Save rsIn
-		for outRowIdx := 0; outRowIdx < rsIn.RowCount; outRowIdx++ {
-			clear(vars)
-			if err := rsIn.ExportToVars(outRowIdx, vars); err != nil {
-				instr.cancelDrainer(fmt.Errorf("cannot export to vars from source table, node %s: %s", node.Name, err.Error()))
-				return bs, instr.waitForDrainer()
-			}
-
-			tableRecord, err = node.TableCreator.CalculateTableRecordFromSrcVars(vars)
-			if err != nil {
-				instr.cancelDrainer(fmt.Errorf("cannot populate table record from [%v], node %s: [%s]", vars, node.Name, err.Error()))
-				return bs, instr.waitForDrainer()
-			}
-
-			// Check table creator having
-			inResult, err = node.TableCreator.CheckTableRecordHavingCondition(tableRecord)
-			if err != nil {
-				instr.cancelDrainer(fmt.Errorf("cannot check having condition [%s], table record [%v], node %s: [%s]", node.TableCreator.RawHaving, tableRecord, node.Name, err.Error()))
-				return bs, instr.waitForDrainer()
-			}
-
-			// Write batch if needed
-			if inResult {
-				err = instr.buildIndexKeys(tableRecord, indexKeyMap)
-				if err != nil {
-					instr.cancelDrainer(fmt.Errorf("cannot build index keys for table %s: [%s]", node.TableCreator.Name, err.Error()))
-					return bs, instr.waitForDrainer()
-				}
-
-				instr.add(tableRecord, indexKeyMap)
-				bs.RowsWritten++
-			}
-		}
-
-		bs.RowsRead += rsIn.RowCount
-
-		// We are tempted to "if rs.RowCount < srcBatchSize break", here but do not do that:
-		// because of the rowid overlapping/epilogue logic, selectBatchFromTableByToken returns less rows than rs capacity
-
-		instr.PCtx.SendHeartbeat()
-	} // for each source table batch
-
-	instr.doneSending()
-	if err := instr.waitForDrainer(); err != nil {
-		return bs, err
-	}
-
-	bs.Elapsed = time.Since(totalStartTime)
-	reportWriteTableComplete(logger, pCtx, bs.RowsRead, bs.RowsWritten, bs.Elapsed, len(node.TableCreator.Indexes), instr.NumWorkers)
-
-	return bs, nil
-}
-
-func RunCreateDistinctTableForBatch(envConfig *env.EnvConfig,
-	logger *l.CapiLogger,
-	pCtx *ctx.MessageProcessingContext,
-	readerNodeRunId int16,
-	startLeftToken int64,
-	endLeftToken int64) (BatchStats, error) {
-
-	logger.PushF("proc.RunCreateDistinctTableForBatch")
-	defer logger.PopF()
-
-	node := pCtx.CurrentScriptNode
-
-	// batchStartTime := time.Now()
-	totalStartTime := time.Now()
-	bs := BatchStats{RowsRead: 0, RowsWritten: 0, Src: node.TableReader.TableName + cql.RunIdSuffix(readerNodeRunId), Dst: node.TableCreator.Name + cql.RunIdSuffix(readerNodeRunId)}
-
-	if readerNodeRunId == 0 {
-		return bs, errors.New("this node has a dependency node to read data from that was never started in this keyspace (readerNodeRunId == 0)")
-	}
-
-	if !node.HasTableReader() {
-		return bs, errors.New("node does not have table reader")
-	}
-	if !node.HasTableCreator() {
-		return bs, errors.New("node does not have table creator")
-	}
-
-	if node.TableCreator.RawHaving != "" {
-		return bs, errors.New("distinct_table node is not allowed to have Having condition")
-	}
-
-	distinctIdxName, _, err := node.TableCreator.GetSingleUniqueIndexDef()
-	if err != nil {
-		return bs, err
-	}
-
-	// Fields to read from source table
-	srcLeftFieldRefs := sc.FieldRefs{}
-	srcLeftFieldRefs.AppendWithFilter(node.TableCreator.UsedInTargetExpressionsFields, sc.ReaderAlias)
-
-	leftBatchSize := node.TableReader.RowsetSize
-	// tableRecordBatchCount := 0
-	curStartLeftToken := startLeftToken
-
-	rsIn := NewRowsetFromFieldRefs(
-		sc.FieldRefs{sc.RowidFieldRef(node.TableReader.TableName)},
-		sc.FieldRefs{sc.RowidTokenFieldRef()},
-		srcLeftFieldRefs)
-
-	instr, err := createInserterAndStartWorkers(logger, envConfig, pCtx, &node.TableCreator, DataIdxSeqModeDistinctIdxFirst, logger.ZapMachine.String)
-	if err != nil {
-		return bs, err
-	}
-	instr.startDrainer()
-	defer instr.closeInserter(logger, pCtx)
-
-	var curStartLeftTokenRowIds []int64
-	for {
-		// Poor man's cache that spans across rsIn retrieved, works well for low-cardinality datasets
-		usedDistinctKeysMap := map[string]struct{}{}
-		distinctCacheHits := 0
-
-		lastRetrievedLeftToken, endTokenRowIds, err := selectBatchFromTableByToken(logger,
-			pCtx,
-			rsIn,
-			node.TableReader.TableName,
-			readerNodeRunId,
-			leftBatchSize,
-			curStartLeftToken,
-			endLeftToken,
-			curStartLeftTokenRowIds)
-		if err != nil {
-			instr.cancelDrainer(fmt.Errorf("cannot select batch from source table, node %s: %s", node.Name, err.Error()))
-			return bs, instr.waitForDrainer()
-		}
-
-		// If token(rowid) guaranteed uniqueness, we would just "curStartLeftToken = lastRetrievedLeftToken + 1"
-		// But duplicates are possible, so we have to be prepared to handle token overlaps
-		// (rows with same token but different rowids returned in separate selectBatchFromTableByToken calls)
-		// See overlap/epilogue logic in selectBatchFromTableByToken.
-		curStartLeftToken = lastRetrievedLeftToken
-		curStartLeftTokenRowIds = endTokenRowIds
-
-		if rsIn.RowCount == 0 {
-			break
-		}
-
-		// Minimize allocations to help GC in this high-traffic loop
-		var tableRecord map[string]any
-		indexKeyMap := map[string]string{}
-		vars := eval.VarValuesMap{}
-
-		// Save rsIn
-		for outRowIdx := 0; outRowIdx < rsIn.RowCount; outRowIdx++ {
-			clear(vars)
-			if err = rsIn.ExportToVars(outRowIdx, vars); err != nil {
-				instr.cancelDrainer(fmt.Errorf("cannot export to vars from source table, node %s: %s", node.Name, err.Error()))
-				return bs, instr.waitForDrainer()
-			}
-
-			tableRecord, err = node.TableCreator.CalculateTableRecordFromSrcVars(vars)
-			if err != nil {
-				instr.cancelDrainer(fmt.Errorf("cannot populate table record from [%v], node %s: [%s]", vars, node.Name, err.Error()))
-				return bs, instr.waitForDrainer()
-			}
-
-			err = instr.buildIndexKeys(tableRecord, indexKeyMap)
-			if err != nil {
-				instr.cancelDrainer(fmt.Errorf("cannot build index keys for table %s: [%s]", node.TableCreator.Name, err.Error()))
-				return bs, instr.waitForDrainer()
-			}
-
-			uniqueDistinctKey := indexKeyMap[distinctIdxName]
-
-			// Poor man's cache
-			if _, ok := usedDistinctKeysMap[uniqueDistinctKey]; !ok {
-				usedDistinctKeysMap[uniqueDistinctKey] = struct{}{}
-				instr.add(tableRecord, indexKeyMap)
-
-				// TODO: this is incorrect, find a way to return proper RowsWritten or abandon them for distinct node altogether
-				bs.RowsWritten++
-			} else {
-				distinctCacheHits++
-				logger.DebugCtx(pCtx, "skipped writing unique key %s to %s, already written", uniqueDistinctKey, node.TableCreator.Name)
-			}
-		}
-
-		logger.DebugCtx(pCtx, "distinct cache hits %d/%d=%d percent %s", distinctCacheHits, rsIn.RowCount, distinctCacheHits*100/rsIn.RowCount, node.TableCreator.Name)
-
-		bs.RowsRead += rsIn.RowCount
-
-		// We are tempted to "if rs.RowCount < srcBatchSize break", here but do not do that:
-		// because of the rowid overlapping/epilogue logic, selectBatchFromTableByToken returns less rows than rs capacity
-
-		instr.PCtx.SendHeartbeat()
-	} // for each source table batch
-
-	instr.doneSending()
-	if err := instr.waitForDrainer(); err != nil {
-		return bs, err
-	}
-
-	bs.Elapsed = time.Since(totalStartTime)
-	reportWriteTableComplete(logger, pCtx, bs.RowsRead, bs.RowsWritten, bs.Elapsed, len(node.TableCreator.Indexes), instr.NumWorkers)
-
-	return bs, nil
-}
 
 func buildKeysToFindInTheLookupIndex(rsLeft *Rowset, scriptNodeLookup sc.LookupDef) ([]string, map[string][]int, error) {
 	// Build keys to find in the lookup index, one key may yield multiple rowids
@@ -891,7 +294,7 @@ func splitKeysIntoChunks(allKeys []string, chunkSize int) [][]string {
 	return keysChunks
 }
 
-func RunCreateTableRelForBatch(envConfig *env.EnvConfig,
+func runCreateTableRelForBatch(envConfig *env.EnvConfig,
 	logger *l.CapiLogger,
 	pCtx *ctx.MessageProcessingContext,
 	readerNodeRunId int16,
@@ -899,7 +302,7 @@ func RunCreateTableRelForBatch(envConfig *env.EnvConfig,
 	startLeftToken int64,
 	endLeftToken int64) (BatchStats, error) {
 
-	logger.PushF("proc.RunCreateTableRelForBatch")
+	logger.PushF("proc.runCreateTableRelForBatch")
 	defer logger.PopF()
 
 	node := pCtx.CurrentScriptNode
